Move registry files to runtime_configuration module
