+++ hpx/plugins/plugin_factory_base.hpp
@@ -9,7 +9,7 @@
 
 #include <hpx/config.hpp>
 #include <hpx/plugin.hpp>
-#include <hpx/plugins/plugin_registry_base.hpp>
+#include <hpx/runtime_configuration/plugin_registry_base.hpp>
 #include <hpx/runtime_configuration/ini.hpp>
 #include <hpx/type_support/pack.hpp>
