+++ libs/runtime_configuration/CMakeLists.txt
@@ -9,18 +9,25 @@ cmake_minimum_required(VERSION 3.3.2 FATAL_ERROR)
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
 set(runtime_configuration_headers
+  hpx/runtime_configuration/component_registry_base.hpp
   hpx/runtime_configuration/find_prefix.hpp
   hpx/runtime_configuration/ini.hpp
+  hpx/runtime_configuration/plugin_registry_base.hpp
+  hpx/runtime_configuration/static_factory_data.hpp
 )
 
 set(runtime_configuration_compat_headers
   hpx/util/find_prefix.hpp
   hpx/util/ini.hpp
+  hpx/plugins/plugin_registry_base.hpp
+  hpx/runtime/components/component_registry_base.hpp
+  hpx/runtime/components/static_factory_data.hpp
 )
 
 set(runtime_configuration_sources
   find_prefix.cpp
   ini.cpp
+  init_ini_data.cpp
 )
 
 include(HPX_AddModule)
