+++ libs/runtime_configuration/include/hpx/runtime_configuration/init_ini_data.hpp
@@ -10,8 +10,8 @@
 
 #include <hpx/filesystem.hpp>
 #include <hpx/plugin.hpp>
-#include <hpx/plugins/plugin_registry_base.hpp>
-#include <hpx/runtime/components/component_registry_base.hpp>
+#include <hpx/runtime_configuration/plugin_registry_base.hpp>
+#include <hpx/runtime_configuration/component_registry_base.hpp>
 #include <hpx/runtime_configuration/ini.hpp>
 
 #include <map>
