+++ libs/runtime_configuration/src/init_ini_data.cpp
@@ -11,12 +11,12 @@
 #include <hpx/filesystem.hpp>
 #include <hpx/logging.hpp>
 #include <hpx/plugin.hpp>
-#include <hpx/plugins/plugin_registry_base.hpp>
-#include <hpx/runtime/components/component_registry_base.hpp>
+#include <hpx/runtime_configuration/plugin_registry_base.hpp>
+#include <hpx/runtime_configuration/component_registry_base.hpp>
 #include <hpx/runtime/startup_function.hpp>
-#include <hpx/util/find_prefix.hpp>
+#include <hpx/runtime_configuration/find_prefix.hpp>
 #include <hpx/runtime_configuration/ini.hpp>
-#include <hpx/util/init_ini_data.hpp>
+#include <hpx/runtime_configuration/init_ini_data.hpp>
 #include <hpx/version.hpp>
 
 #include <boost/assign/std/vector.hpp>
