+++ libs/runtime_configuration/tests/CMakeLists.txt
@@ -36,7 +36,6 @@ if (HPX_RUNTIME_CONFIGURATION_WITH_TESTS)
       modules.runtime_configuration
       HEADERS ${runtime_configuration_headers}
       HEADER_ROOT ${PROJECT_SOURCE_DIR}/include
-      NOLIBS
       DEPENDENCIES hpx_runtime_configuration)
   endif()
 endif()
